# api/openapi.policy.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/admin/policy-corrections" : {
      "get" : {
        "operationId" : "listPolicyCorrections",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "maximum" : 1000,
            "minimum" : 1
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20,
            "maximum" : 50,
            "minimum" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSlicePolicyCorrectionItem"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "입력 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "보정 처리 또는 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "관리자 정책 보정·충돌·해제 이력",
        "tags" : [ "policy-correction-controller" ]
      },
      "post" : {
        "description" : "정책당 한 항목을 보정하며 원본을 유지한다. 같은 요청 ID·입력·작업자는 기존 결과를 반환한다.",
        "operationId" : "createPolicyCorrection",
        "parameters" : [ {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PolicyCorrectionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyCorrectionItem"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "입력 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "개정 변경·진행 중 보정·요청 ID 충돌"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "보정 처리 또는 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "공개 정책의 정책명 또는 운영 기관 보정",
        "tags" : [ "policy-correction-controller" ]
      }
    },
    "/api/v1/admin/policy-corrections/policies/{number}" : {
      "get" : {
        "operationId" : "getCorrectionPolicy",
        "parameters" : [ {
          "in" : "path",
          "name" : "number",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "[0-9]{1,100}"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionExceptionCurrentPolicy"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "입력 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "공개 정책 없음"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "보정 처리 또는 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "보정할 공개 정책의 현재 내용 조회",
        "tags" : [ "policy-correction-controller" ]
      }
    },
    "/api/v1/admin/policy-corrections/{id}/resolutions" : {
      "post" : {
        "description" : "조회한 현재 개정과 검토 원본 ID가 같을 때만 처리한다. KEEP은 충돌 상태에서만 허용한다. 처리 후 남은 수집 실패 항목은 별도로 재처리한다.",
        "operationId" : "resolvePolicyCorrection",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "description" : "GET /api/v1/session에서 받은 csrfToken",
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PolicyCorrectionResolution"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyCorrectionItem"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "입력 오류"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "로그인 필요"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "권한 또는 CSRF 토큰 확인 필요"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "개정·검토 원본 변경 또는 요청 ID 충돌"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "보정 처리 또는 조회 실패"
          }
        },
        "security" : [ {
          "memberSession" : [ ]
        } ],
        "summary" : "정책 보정 해제 또는 새 원본 기준 보정 유지",
        "tags" : [ "policy-correction-controller" ]
      }
    },
...
    "/api/v1/policies/{policyNumber}" : {
      "get" : {
        "operationId" : "getPolicy",
        "parameters" : [ {
          "in" : "path",
          "name" : "policyNumber",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyDetailResponse"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "503" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PolicyApiError"
                }
              }
            },
            "description" : "정책 저장소 조회 실패"
          }
        },
        "summary" : "정책 상세와 원문 출처 조회",
        "tags" : [ "policy-catalog-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AdminSlicePolicyCorrectionItem" : {
        "type" : "object",
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyCorrectionItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "1부터 시작하는 페이지"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "hasNext", "items", "page", "pageSize" ]
      },
...
      "CollectionExceptionCurrentPolicy" : {
        "type" : "object",
        "properties" : {
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "correctionId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "previousRevision" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CollectionExceptionRevision"
            }, {
              "type" : "null"
            } ],
            "description" : "같은 정책의 직전 내부 개정. 없으면 null"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sourceCapturedAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "현재 개정이 참조하는 원본 수집 시각. 개정 적용 시각이 아님"
          }
        },
        "required" : [ "collectedAt", "content", "correctionId", "policyNumber", "previousRevision", "revision", "sourceCapturedAt" ]
      },
...
      "CollectionExceptionRevision" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "correctionId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "서비스 내부 개정 번호"
          },
          "sourceCapturedAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "개정이 참조하는 원본 수집 시각. 개정 적용 시각이 아님"
          }
        },
        "required" : [ "content", "correctionId", "revision", "sourceCapturedAt" ]
      },
...
      "PolicyApiError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      },
...
      "PolicyContent" : {
        "type" : "object",
        "properties" : {
          "applicationPeriod" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "links" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyOfficialLink"
            }
          },
          "organization" : {
            "type" : "string"
          },
          "regionCodes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "sections" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PolicyTextSection"
            }
          },
          "sourceModifiedAtText" : {
            "type" : "string",
            "description" : "시간대가 확정되지 않은 원천 수정 일시 원문"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "applicationPeriod", "category", "description", "links", "organization", "regionCodes", "sections", "sourceModifiedAtText", "title" ]
      },
      "PolicyCorrectionItem" : {
        "type" : "object",
        "properties" : {
          "actorId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "appliedRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "currentRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "field" : {
            "type" : "string",
            "enum" : [ "TITLE", "ORGANIZATION" ]
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "requestedRevision" : {
            "type" : "integer",
            "format" : "int64"
          },
          "resolution" : {
            "type" : [ "string", "null" ]
          },
          "resolvedAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "resolvedBy" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "resolvedReason" : {
            "type" : [ "string", "null" ]
          },
          "resolvedRevision" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "reviewContent" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "reviewSnapshotId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "reviewValue" : {
            "type" : "string"
          },
          "sourceValue" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "CONFLICT", "RELEASED" ]
          },
          "value" : {
            "type" : "string"
          }
        },
        "required" : [ "actorId", "appliedRevision", "createdAt", "currentRevision", "field", "id", "policyNumber", "reason", "requestedRevision", "resolution", "resolvedAt", "resolvedBy", "resolvedReason", "resolvedRevision", "reviewContent", "reviewSnapshotId", "reviewValue", "sourceValue", "status", "value" ]
      },
      "PolicyCorrectionRequest" : {
        "type" : "object",
        "properties" : {
          "expectedRevision" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "field" : {
            "type" : "string",
            "enum" : [ "TITLE", "ORGANIZATION" ]
          },
          "policyNumber" : {
            "type" : "string",
            "pattern" : "[0-9]{1,100}"
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "value" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "expectedRevision", "field", "policyNumber", "reason", "requestId", "value" ]
      },
      "PolicyCorrectionResolution" : {
        "type" : "object",
        "properties" : {
          "action" : {
            "type" : "string",
            "enum" : [ "KEEP", "USE_SOURCE" ]
          },
          "expectedRevision" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          },
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "requestId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "reviewSnapshotId" : {
            "type" : "integer",
            "format" : "int64",
            "minimum" : 0
          }
        },
        "required" : [ "action", "expectedRevision", "reason", "requestId", "reviewSnapshotId" ]
      },
...
      "PolicyDetailResponse" : {
        "type" : "object",
        "properties" : {
          "collectedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "content" : {
            "$ref" : "#/components/schemas/PolicyContent"
          },
          "policyNumber" : {
            "type" : "string"
          },
          "recruitment" : {
            "$ref" : "#/components/schemas/PolicyRecruitment"
          },
          "revision" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "서비스 내부 적용 개정 번호"
          },
          "sourceConditions" : {
            "type" : "array",
            "description" : "온통청년에 등록된 조건 표기. 자격 판정이 아니며 제한을 표기한 값만 담는다. 빈 배열은 조건이 없다는 뜻이 아니다.",
            "items" : {
              "$ref" : "#/components/schemas/PolicySourceCondition"
            }
          },
          "sourceNotices" : {
            "type" : "array",
            "description" : "현재 수집 내용에서 확인한 조건 충돌 안내. 빈 배열은 검토 완료를 뜻하지 않는다.",
            "items" : {
              "$ref" : "#/components/schemas/PolicySourceNotice"
            }
          },
          "sourceUrl" : {
            "type" : "string"
          }
        },
        "required" : [ "collectedAt", "content", "policyNumber", "recruitment", "revision", "sourceConditions", "sourceNotices", "sourceUrl" ]
      },
...
      "PolicyOfficialLink" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "required" : [ "label", "url" ]
      },
...
      "PolicyRecruitment" : {
        "type" : "object",
        "properties" : {
          "daysUntilDeadline" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "평가 시점의 서울 날짜부터 마감일까지 남은 일수. 마감일이 없으면 null"
          },
          "deadlineOnSeoul" : {
            "type" : [ "string", "null" ],
            "format" : "date",
            "description" : "확인한 마감일(서울). 상시·소진형·기간 미확인과 날짜 없는 마감은 null"
          },
          "evaluatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "explanation" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "BEFORE_OPENING", "OPEN", "CLOSED", "ROLLING", "UNTIL_EXHAUSTED", "UNKNOWN" ]
          }
        },
        "required" : [ "daysUntilDeadline", "deadlineOnSeoul", "evaluatedAt", "explanation", "status" ]
      },
...
      "PolicySourceCondition" : {
        "type" : "object",
        "properties" : {
          "label" : {
            "type" : "string",
            "description" : "조건 이름. 연령·소득·취업 상태·학력"
          },
          "value" : {
            "type" : "string",
            "description" : "온통청년에 표기된 값"
          }
        },
        "required" : [ "label", "value" ]
      },
      "PolicySourceNotice" : {
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "sourceLabel" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "description", "sourceLabel", "sourceUrl", "title" ]
      },
...
      "PolicyTextSection" : {
        "type" : "object",
        "properties" : {
          "text" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "text", "title" ]
      },
...
    },
    "securitySchemes" : {
      "memberSession" : {
        "description" : "카카오·네이버 로그인 후 발급하는 서버 세션",
        "in" : "cookie",
        "name" : "YPM_SESSION",
        "type" : "apiKey"
      }
    }
  }
}